misc: dev env improvements
This patch enables get_feature_gates which was missing in kustomize.sh

It also adds a make rebuild-vpp target in vpp-manager to do a local
rebuild of VPP whithout the debs to shorten build iterations in dev

This also adds more context to the errors returned by the net_watcher.

+++ vpp-manager/images/ubuntu-build/build_script.sh
@@ -5,6 +5,8 @@ set -o errexit
 cd ${VPP_MGR_DIR}/vpp_build
 
 make build-release
-rm -f ./build-root/*.deb ./build-root/*.changes ./build-root/*.buildinfo
-make pkg-deb
+if [ "${NO_BUILD_DEBS}" != "true" ]; then
+	rm -f ./build-root/*.deb ./build-root/*.changes ./build-root/*.buildinfo
+	make pkg-deb
+fi
